% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rf_to_reach_final.R
\name{calculate_rf_to_reach_final}
\alias{calculate_rf_to_reach_final}
\title{Calculate Raising Factors for Conversion to Nominal Data}
\usage{
calculate_rf_to_reach_final(df_to_reach, df_to_convert, strata_cols)
}
\arguments{
\item{strata_cols}{A character vector of column names used for stratifying the data (e.g., \code{species}, \code{gear_type}, \code{fishing_fleet}, \code{year}).}

\item{nominal_df}{A dataframe containing nominal data with columns for measurement units and measurement values.}

\item{lvl0_conv_df}{A dataframe containing georeferenced data with both "number" and "tons" units, from which conversion factors will be calculated.}
}
\value{
A dataframe with calculated raising factors for each stratum where conversion is applicable. If conversion factors are too high,
they are marked as inconclusive. For very low conversion factors, potential inconsistencies are indicated.
}
\description{
This function calculates the average conversion factors needed to match georeferenced data in numbers with nominal data in tons.
It only considers strata where both measurement units (numbers and tons) are available and evaluates the coherence of conversion factors
to achieve 100\% of the nominal data for the same strata. If conversion factors are excessively high, the results are inconclusive,
as nominal data should generally exceed georeferenced data; here, both conversion and raising occur. However, if the conversion factors
are very low, an inconsistency between the datasets is likely.
}
\details{
The function groups and aggregates data within each specified stratum. It calculates the conversion factors for each stratum
by determining the difference between nominal data values and converted georeferenced data values in tons. Strata with both measurement
units are selected for comparison.
}
\examples{
\dontrun{
# Example usage:
nominal_data <- data.frame(time_start = as.Date("2020-01-01"), measurement_value = 100, ...)
georef_data <- data.frame(time_start = as.Date("2020-01-01"), measurement_unit = "no", measurement_value = 200, ...)
strata_cols <- c("species", "gear_type", "fishing_fleet", "year", "geographic_identifier_nom")
calculate_rf_to_reach_final.R(nominal_df = nominal_data, lvl0_conv_df = georef_data, strata_cols = strata_cols)
}

}
