% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dimension_differences.R
\name{compare_dimension_differences}
\alias{compare_dimension_differences}
\title{Calculate and Visualize Differences for Each Dimension}
\usage{
compare_dimension_differences(
  Groupped_all,
  Other_dimensions,
  parameter_diff_value_or_percent,
  parameter_columns_to_keep = c("Precision", "measurement_unit", "Values dataset 1",
    "Values dataset 2", "Loss / Gain", "Difference (in \%)", "Dimension",
    "Difference in value"),
  topn = 6,
  outputonly = FALSE
)
}
\arguments{
\item{Groupped_all}{Data frame containing grouped strata data.}

\item{Other_dimensions}{Vector of dimensions to be analyzed.}

\item{parameter_diff_value_or_percent}{Character string indicating the parameter to sort by ("Difference in value" or "Difference (in \%)").}

\item{parameter_columns_to_keep}{Vector of column names to keep in the final output.}

\item{topn}{Integer indicating the number of top differences to display.}

\item{outputonly}{Logical value indicating whether to output only the image.}
}
\value{
A data frame containing the summarized differences for each dimension.
}
\description{
This function calculates the differences in various dimensions between two datasets
and provides a detailed summary of the differences.
}
\examples{
\dontrun{
compare_dimension_differences(Groupped_all, c("Dimension1", "Dimension2"), "Difference in value", c("Column1", "Column2"), 6, FALSE)
}
}
\author{
Bastien Grasset, \email{bastien.grasset@ird.fr}
}
