% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_georef_nominal.R
\name{compare_georef_nominal}
\alias{compare_georef_nominal}
\title{Compare Georef and Nominal Data}
\usage{
compare_georef_nominal(georef, nominal, connectionDB)
}
\arguments{
\item{georef}{Data frame containing georeferenced data.}

\item{nominal}{Data frame containing nominal data.}

\item{connectionDB}{Database connection object.}
}
\value{
A list containing the comparison results.
}
\description{
This function takes georef and nominal data as inputs, performs a comparison, and returns the results.
}
