% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_strata_differences.R
\name{compare_strata_differences}
\alias{compare_strata_differences}
\title{Calculate and Compare Strata Differences}
\usage{
compare_strata_differences(
  init,
  final,
  Groupped_all,
  titre_1 = "Dataset1",
  titre_2 = "Dataset2",
  parameter_columns_to_keep,
  unique_analyse = FALSE
)
}
\arguments{
\item{init}{Data frame containing initial geographical data.}

\item{final}{Data frame containing final geographical data.}

\item{Groupped_all}{Data frame containing grouped strata data.}

\item{titre_1}{Title for the first dataset.}

\item{titre_2}{Title for the second dataset.}

\item{parameter_columns_to_keep}{Vector of column names to keep in the final output.}

\item{unique_analyse}{Logical value indicating whether the analysis is unique.}
}
\value{
A list containing summaries and visualizations of the strata differences.
}
\description{
This function calculates the differences in strata between two datasets and
provides various summaries and visualizations of the results.
}
\examples{
\dontrun{
compare_strata_differences(init, final, Groupped_all, "Dataset1", "Dataset2", c("Column1", "Column2"), FALSE)
}
}
