% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compnumberstratas.R
\name{compnumberstratas}
\alias{compnumberstratas}
\title{Perform Strata Comparisons}
\usage{
compnumberstratas(
  init,
  final,
  Groupped_all,
  titre_1,
  titre_2,
  parameter_columns_to_keep,
  unique_analyse = FALSE
)
}
\arguments{
\item{init}{A data frame representing the initial dataset.}

\item{final}{A data frame representing the final dataset.}

\item{Groupped_all}{A data frame containing grouped data.}

\item{titre_1}{A string representing the title for the initial dataset.}

\item{titre_2}{A string representing the title for the final dataset.}

\item{parameter_columns_to_keep}{A vector of column names to keep in the final comparison.}

\item{unique_analyse}{A boolean flag indicating whether to perform a unique analysis. Defaults to FALSE.}
}
\value{
A list containing the following elements:
\item{strates_perdues_first_10}{A data frame of the first 10 lost strata.}
\item{number_init_column_final_column}{A data frame comparing the number of unique values in columns between the initial and final datasets.}
\item{disapandap}{A data frame showing the differences in strata.}
}
\description{
This function performs various comparisons between initial and final datasets to identify lost and gained strata.
}
\examples{
\dontrun{
compnumberstratas(init, final, Groupped_all, titre_1, titre_2, parameter_columns_to_keep, unique_analyse = FALSE)
}
}
\author{
Bastien Grasset, \email{bastien.grasset@ird.fr}
}
