% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{download_codelists}
\alias{download_codelists}
\title{download_codelists}
\format{
An object of class \code{list} of length 9.
}
\usage{
download_codelists
}
\description{
Liste des codelists FDIWG, chaque élément est un data frame contenant au moins
les colonnes \code{code} et \code{label} :
}
\details{
\describe{
\item{cl_asfis_species}{Codelist des espèces selon la classification ASFIS}
\item{cl_measurement_processing_level}{Niveaux de traitement des mesures}
\item{cl_measurement}{Types de mesures FDI}
\item{cl_fishing_mode}{Modes de pêche (GTA FIRMS)}
\item{cl_isscfg_pilot_gear}{Engins pilotes ISSCFG (GTA FIRMS)}
\item{cl_fishingfleet_firms}{Flottes de pêche (GTA FIRMS)}
\item{cl_catch_concepts}{Concepts relatifs aux captures (CWP FDIWG)}
\item{cl_measurement_types_effort}{Types de mesures d’effort FDI}
\item{cl_areal_grid}{Quadrillage spatial (CWP FDIWG)}
}
}
\keyword{datasets}
