% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_function.R, R/isnulllist.R
\name{filtering_function}
\alias{filtering_function}
\title{Filter Dataframe based on provided Parameters /Dimensions}
\usage{
filtering_function(dataframe_to_filter, parameter_filtering)

filtering_function(dataframe_to_filter, parameter_filtering)
}
\arguments{
\item{dataframe_to_filter}{A data frame to be filtered.}

\item{parameter_filtering}{A list of parameters to use for filtering.}
}
\value{
A filtered data frame.

A filtered data frame.
}
\description{
This function filters a data frame based on specified filtering parameters.

This function filters a data frame based on specified filtering parameters.
}
