% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R
\name{fonction_groupement}
\alias{fonction_groupement}
\title{Group and Summarize Data}
\usage{
fonction_groupement(these_col, init, final)
}
\arguments{
\item{these_col}{A character vector of column names to group by.}

\item{init}{A data.table containing the initial measurement data.}

\item{final}{A data.table containing the final measurement data.}
}
\value{
A data.table containing the results of the comparison between the
two input data.tables, including summed values, losses or gains,
and percentage differences.
}
\description{
This function takes two data.tables and groups them by specified columns,
summing the measurement values for each group, and then compares the results
from both data.tables. It computes the loss or gain in measurement values
and provides additional metrics related to the comparison.
}
