% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/function_geographic_identifier_renaming_and_not_standards_unit.R
\name{function_geographic_identifier_renaming_and_not_standards_unit}
\alias{function_geographic_identifier_renaming_and_not_standards_unit}
\title{Rename Geographic Identifiers and Handle Standard Units}
\usage{
function_geographic_identifier_renaming_and_not_standards_unit(
  dataframe_to_filter,
  geo_dim,
  parameter_fact,
  geo_dim_group
)
}
\arguments{
\item{dataframe_to_filter}{A data frame to filter and rename columns.}

\item{geo_dim}{The geographic dimension to rename.}

\item{parameter_fact}{A parameter indicating the measurement context.}

\item{geo_dim_group}{The grouping geographic dimension to rename.}
}
\value{
A modified data frame with renamed columns.
}
\description{
This function renames geographic identifiers in a data frame and manages non-standard units.
}
