% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_recap_each_step.R
\name{function_recap_each_step}
\alias{function_recap_each_step}
\title{Function to Record Step Details and Save Results}
\usage{
function_recap_each_step(
  step_name,
  rds_data,
  explanation = "No explanation provided to this step",
  functions = "No function used in this step",
  option_list = NULL,
  entity = NULL
)
}
\arguments{
\item{step_name}{A character string specifying the name of the step.}

\item{rds_data}{A data frame containing the data to be saved as an RDS file.}

\item{explanation}{A character string providing an explanation of the step.}

\item{functions}{A character string listing the functions used in the step.}

\item{option_list}{A list of options used in the step.}

\item{entity}{A geoflow entity}
}
\value{
None
}
\description{
This function records the details of each processing step, including explanations, function names, and options. It saves the results as RDS files and text files in a specified directory.
}
\examples{
\dontrun{
function_recap_each_step("step1", data, "This step does X", "function1, function2", list(option1 = "value1"))
}
}
\author{
Bastien Grasset
}
