% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bookdown_yml.R
\name{generate_bookdown_yml}
\alias{generate_bookdown_yml}
\title{Generate a \verb{_bookdown.yml} file dynamically}
\usage{
generate_bookdown_yml(
  destination = system.file("rmd", package = "CWP.dataset"),
  new_session = FALSE
)
}
\arguments{
\item{destination}{Character. The directory where the \verb{_bookdown.yml} file
will be created. Default is \code{"bookdown_run"}.}

\item{new_session}{Boolean. Is the bookdown yml with new_session parameter to true or false}
}
\value{
No return value. The function writes \verb{_bookdown.yml} in the
specified destination directory.
}
\description{
This function creates a \verb{_bookdown.yml} file in the specified directory
to define the structure and configuration for a Bookdown project.
It allows specifying the output directory, merging behavior, and the list
of \code{.Rmd} files required for the Bookdown compilation.
}
\examples{
# Generate a `_bookdown.yml` in a custom directory
generate_bookdown_yml("custom_bookdown_dir")

}
