% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qflextable2.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate \code{\%in\%} Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{A vector of elements to check.}

\item{table}{A vector to check against.}
}
\value{
Logical vector indicating if each element of \code{x} is not in \code{table}.
}
\description{
This operator checks if elements are not present in a given vector.
}
\examples{
5 \%notin\% c(1, 2, 3) # TRUE
"a" \%notin\% c("b", "c") # TRUE
2 \%notin\% c(1, 2, 3) # FALSE

}
