% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groupping_differences.R
\name{groupping_differences}
\alias{groupping_differences}
\title{Groupping Differences}
\usage{
groupping_differences(
  init,
  final,
  parameter_time_dimension,
  parameter_geographical_dimension,
  parameter_geographical_dimension_groupping
)
}
\arguments{
\item{init}{Initial dataset.}

\item{final}{Final dataset.}

\item{parameter_time_dimension}{Vector of time dimensions.}

\item{parameter_geographical_dimension}{Vector of geographical dimensions.}

\item{parameter_geographical_dimension_groupping}{Vector of geographical dimension groupings.}
}
\value{
A list containing grouped differences for all dimensions, geographical dimension groupings, and time dimensions.
}
\description{
This function groups differences between the initial and final datasets based on various dimensions.
}
\examples{
# Example usage:
results <- groupping_differences(init, final, parameter_time_dimension, parameter_geographical_dimension, parameter_geographical_dimension_groupping)
}
