% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_default.R
\name{is_ggplot}
\alias{is_ggplot}
\title{Check if an Object is a ggplot}
\usage{
is_ggplot(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
Logical value indicating whether the object is a ggplot.
}
\description{
This function checks if an object is a ggplot object.
}
