% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_null_or_not_exist.R
\name{is_null_or_not_exist}
\alias{is_null_or_not_exist}
\title{Check if Variable is NULL or Does Not Exist}
\usage{
is_null_or_not_exist(x)
}
\arguments{
\item{x}{The variable to check.}
}
\value{
Logical value indicating whether the variable is NULL or does not exist.
}
\description{
This function checks if a variable is either NULL or does not exist in the environment.
}
