% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnulllist.R
\name{isNullList}
\alias{isNullList}
\title{Check if All Elements in a List are NULL or Empty}
\usage{
isNullList(x)
}
\arguments{
\item{x}{A list to check.}
}
\value{
Logical value indicating whether all elements are NULL or empty.
}
\description{
This function checks if all elements in a list are of zero length.
}
