% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_plots_subfigures.R
\name{knitting_plots_subfigures}
\alias{knitting_plots_subfigures}
\title{Save Plots with Subfigures in a Knitr Environment}
\usage{
knitting_plots_subfigures(
  plot,
  title,
  folder = "Unknown_folder",
  fig.pathinside = fig.path
)
}
\arguments{
\item{plot}{The plot object to save.}

\item{title}{A character string for the plot title.}

\item{folder}{The folder where the plot will be saved.}

\item{fig.pathinside}{The path for saving the figure.}
}
\value{
None
}
\description{
This function saves plots and creates subfigures for rendering in RMarkdown.
}
