% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_path.R
\name{last_path}
\alias{last_path}
\title{Extract Last Part of a File Path}
\usage{
last_path(x)
}
\arguments{
\item{x}{A character string representing the file path.}
}
\value{
A character string representing the last part of the file path.
}
\description{
This function returns the last part of a file path after removing a specific suffix.
}
