% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_path_reduced.R
\name{last_path_reduced}
\alias{last_path_reduced}
\title{Reduce Last Path for Georeferenced Datasets}
\usage{
last_path_reduced(x)
}
\arguments{
\item{x}{A character string representing the file path.}
}
\value{
A character string representing the modified last path.
}
\description{
This function modifies the last path by removing specific substrings for clarity.
}
