% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_dimension_analysis.R
\name{other_dimension_analysis}
\alias{other_dimension_analysis}
\title{Calculate and Visualize Data Distribution for Other Dimensions}
\usage{
other_dimension_analysis(
  Other_dimensions,
  init,
  final,
  titre_1,
  titre_2,
  unique_analyse = FALSE,
  fig.path,
  topn = 7,
  deferred = FALSE
)
}
\arguments{
\item{Other_dimensions}{A vector of dimensions to analyze.}

\item{init}{Initial dataset.}

\item{final}{Final dataset.}

\item{titre_1}{Title for the first dataset.}

\item{titre_2}{Title for the second dataset.}

\item{unique_analyse}{Logical indicating whether the analysis is unique.}

\item{fig.path}{Path to save the figures.}

\item{topn}{An integer for the number of top categories to display.}

\item{deferred}{Logical. If \code{TRUE}, the charts are returned as deferred plots (their description,
to be drawn with \code{\link[=cwp_materialise_plot]{cwp_materialise_plot()}}) instead of plot objects. Default \code{FALSE}.}
}
\value{
A list containing the pie charts and bar charts for each dimension.
}
\description{
This function calculates and visualizes the data distribution for various dimensions
using pie charts and bar charts.
}
\examples{
\dontrun{
other_dimension_analysis(c("Dimension1", "Dimension2"), init, final, "Dataset1", "Dataset2", FALSE, "path/to/save")
}
}
\author{
Bastien Grasset, \email{bastien.grasset@ird.fr}
}
