% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_dimension_analysis_dygraphs.R
\name{other_dimension_analysis_dygraphs}
\alias{other_dimension_analysis_dygraphs}
\title{Calculate and Visualize Data Distribution for Various Dimensions Using dygraphs}
\usage{
other_dimension_analysis_dygraphs(
  Other_dimensions,
  init,
  final = NULL,
  titre_1 = "Dataset1",
  titre_2 = "Dataset2",
  unique_analyse = FALSE,
  fig.path = NULL,
  topn = 7
)
}
\arguments{
\item{Other_dimensions}{A vector of dimensions to analyze.}

\item{init}{Initial dataset.}

\item{final}{Final dataset (optional).}

\item{titre_1}{Title for the first dataset.}

\item{titre_2}{Title for the second dataset (optional).}

\item{unique_analyse}{Logical indicating whether the analysis is unique.}

\item{fig.path}{Path to save the figures (optional).}

\item{topn}{The number of top categories to include in the chart.}
}
\value{
A list containing the dygraphs for each dimension.
}
\description{
This function calculates and visualizes the data distribution for various dimensions
using interactive time-series plots with dygraphs for two datasets or a single dataset.
}
\examples{
\dontrun{
data_dimension_analysis_dygraphs(c("Dimension1", "Dimension2"), init, final, "Dataset1", "Dataset2", FALSE)
}
}
