% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{parameter_final_catch}
\alias{parameter_final_catch}
\title{Example subset of catch dataset (final)}
\format{
Same structure as \code{parameter_init_catch}.
}
\usage{
parameter_final_catch
}
\description{
A small sample of final catch data, structured according to CWP standards.
}
\keyword{datasets}
