% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{parameter_final_effort}
\alias{parameter_final_effort}
\title{Example subset of effort dataset (final)}
\format{
Same structure as \code{parameter_init_effort}.
}
\usage{
parameter_final_effort
}
\description{
A small sample of final effort data, structured according to CWP standards.
}
\keyword{datasets}
