% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{parameter_init_catch}
\alias{parameter_init_catch}
\title{Example subset of catch dataset (initial)}
\format{
A tibble with 200 rows and 14 variables:
\describe{
\item{source_authority}{Source of the data (e.g. RFMO)}
\item{gear_type}{CWP gear type code}
\item{fishing_fleet}{Fishing fleet country or code}
\item{fishing_mode}{Fishing mode (e.g., UNK for unknown)}
\item{time_start}{Start date of the operation}
\item{time_end}{End date of the operation}
\item{geographic_identifier}{CWP geographic identifier}
\item{measurement_unit}{Unit of measurement (e.g., t, no, etc.)}
\item{measurement_value}{Measurement value (e.g., catch in tons)}
\item{GRIDTYPE}{Spatial resolution label}
\item{Gear}{Gear label}
\item{fishing_mode_label}{Fishing mode label}
\item{fishing_fleet_label}{Fishing fleet label}
\item{species}{Species code (e.g., YFT, SKJ, etc.)}
}
}
\usage{
parameter_init_catch
}
\description{
A small sample of initial catch data, structured according to CWP standards.
}
\keyword{datasets}
