% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_default.R, R/qflextable2.R
\name{qflextable2}
\alias{qflextable2}
\title{Create and Save a Flextable}
\usage{
qflextable2(
  x,
  captionn = NULL,
  autonumm = autonum,
  pgwidth = 6,
  columns_to_color = NULL,
  save_folder = NULL,
  fig.pathinside = "Figures",
  grouped_data = NULL,
  interactive_plot = FALSE,
  find_and_print = FALSE
)

qflextable2(
  x,
  captionn = NULL,
  autonumm = autonum,
  pgwidth = 6,
  columns_to_color = NULL,
  save_folder = NULL,
  fig.pathinside = "Figures",
  grouped_data = NULL,
  interactive_plot = FALSE,
  find_and_print = FALSE
)
}
\arguments{
\item{x}{A data frame or flextable to create the table from.}

\item{captionn}{An optional character string for the table caption.}

\item{autonumm}{An optional automatic numbering parameter.}

\item{pgwidth}{A numeric value for the width of the table.}

\item{columns_to_color}{Optional columns to apply color coding.}

\item{save_folder}{Optional folder to save the flextable.}

\item{fig.pathinside}{A character string for the path to save figures.}

\item{grouped_data}{Optional grouped data for formatting.}

\item{interactive_plot}{Logical indicating if the output should be interactive.}
}
\value{
A flextable object or a DT datatable if interactive_plot is TRUE.

A flextable object or a DT datatable if interactive_plot is TRUE.
}
\description{
This function creates a flextable from a data frame, optionally saving it as an image.

This function creates a flextable from a data frame, optionally saving it as an image.
}
