% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Data from Various File Types}
\usage{
read_data(file_path)
}
\arguments{
\item{file_path}{A character string indicating the path of the file to read.}
}
\value{
Data read from the specified file, either as a data frame or data table.
}
\description{
This function reads data based on the provided file path and file type.
}
\examples{
# Example: Reading a CSV file
csv_data <- read_data(here::here("inst/CWP_dataset.csv"))
head(csv_data)  # Display the first few rows
}
