% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_plots_subfigures.R
\name{render_subfigures}
\alias{render_subfigures}
\title{Render Subfigures from a List of Plots}
\usage{
render_subfigures(plots_list, titles_list, general_title)
}
\arguments{
\item{plots_list}{A list of plot objects.}

\item{titles_list}{A list of character strings for plot titles.}

\item{general_title}{A character string for the general title of the plots.}
}
\value{
None
}
\description{
This function creates a grid of subfigures from a list of plots and titles.
}
