% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_groupement.R,
%   R/separate_chunks_and_text.R
\name{separate_chunks_and_text}
\alias{separate_chunks_and_text}
\title{Separate R Code Chunks and Text from an RMarkdown (.Rmd) File and Optionally Save Them}
\usage{
separate_chunks_and_text(rmd_file, save_files = FALSE)

separate_chunks_and_text(rmd_file, save_files = FALSE)
}
\arguments{
\item{rmd_file}{A character string specifying the path to the .Rmd file.}

\item{save_files}{Logical. If TRUE, saves the chunks to an \code{.R} file and the text to a \code{.txt} file.
Default is FALSE.}
}
\value{
A list containing two elements:
\item{chunks}{A character vector with the content of all R code chunks.}
\item{text}{A character vector with the content outside of R code chunks.}

A list containing two elements:
\item{chunks}{A character vector with the content of all R code chunks.}
\item{text}{A character vector with the content outside of R code chunks.}
}
\description{
This function reads an RMarkdown file and separates the code chunks from the text.
It returns a list with two elements: \code{chunks} (R code chunks) and \code{text} (non-code content).
Optionally, it saves the chunks to an \code{.R} file and the text to a \code{.txt} file.

This function reads an RMarkdown file and separates the code chunks from the text.
It returns a list with two elements: \code{chunks} (R code chunks) and \code{text} (non-code content).
Optionally, it saves the chunks to an \code{.R} file and the text to a \code{.txt} file.
}
