% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidying_data.R
\name{tidying_data}
\alias{tidying_data}
\title{Tidying Data by Keeping Specific Columns}
\usage{
tidying_data(dataframe, parameter_colnames_to_keep_dataframe, time_dimension)
}
\arguments{
\item{dataframe}{A data frame to tidy.}

\item{parameter_colnames_to_keep_dataframe}{A character vector of column names to keep.}

\item{time_dimension}{A character vector of time dimension column names.}
}
\value{
A tidied data frame.
}
\description{
This function tidies a data frame by selecting specified columns and converting time columns to character.
}
