% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecoverage.R
\name{timecoverage}
\alias{timecoverage}
\title{Perform Time Coverage Analysis}
\usage{
timecoverage(
  parameter_time_dimension,
  unique_analyse,
  titre_1 = "titre_1",
  titre_2 = "titre_2",
  time_dimension_list_groupped,
  fig.path = "figures"
)
}
\arguments{
\item{parameter_time_dimension}{A string representing the time dimension.}

\item{unique_analyse}{A boolean flag indicating whether to perform a unique analysis. Defaults to FALSE.}

\item{titre_1}{A string representing the title for the initial dataset.}

\item{titre_2}{A string representing the title for the final dataset.}

\item{time_dimension_list_groupped}{A list of data frames representing the grouped time dimensions.}

\item{fig.path}{A string representing the path to save the figures.}
}
\value{
A list containing the following elements:
\item{time_dimension_list_groupped_diff_image_knit}{A list of knitted plots for time dimension differences.}
}
\description{
This function analyzes the time coverage of values for specified time dimensions and generates plots of the differences.
}
\examples{
timecoverage(unique_analyse, parameter_time_dimension, titre_1, titre_2, time_dimension_list_groupped, fig.path)
}
