% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_options_to_csv.R
\name{write_options_to_csv}
\alias{write_options_to_csv}
\title{Write options to CSV and return structured output}
\usage{
write_options_to_csv(opts)
}
\arguments{
\item{opts}{A named list where each element is a vector of options corresponding to an entity.}
}
\value{
A list containing:
\item{options}{A named list where each key is \verb{options_<name>} and the value is a concatenated string of options.}
\item{table}{A data frame with two columns: "Options" (names of the options) and "Position" (concatenated values).}
}
\description{
This function writes a CSV file containing all the options of the data column of the entity of geoflow for GTA workflows
Additionally, it returns a structured list containing the options renamed as a named list.
}
