% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_book.R
\name{build_book}
\alias{build_book}
\title{Build a bookdown project in isolated sessions}
\usage{
build_book(
  master_qs_rel,
  orig_setup_rmd = system.file("rmd/Setup_markdown.Rmd", package = "CWP.dataset"),
  src_paths = c(system.file("rmd/index.Rmd", package = "CWP.dataset"),
    system.file("rmd/first_and_first_to_last_and_process.Rmd", package = "CWP.dataset"),
    system.file("rmd/all_child_process.Rmd", package = "CWP.dataset"),
    system.file("rmd/Annexenew.Rmd", package = "CWP.dataset")),
  book_filename = "tableau_recapbookdowntest",
  output_dir = "_book",
  template = system.file("rmd/template.tex", package = "CWP.dataset"),
  output_format = "bookdown::pdf_document2",
  new_session = TRUE,
  delete_merged_file = TRUE,
  root = here::here()
)
}
\arguments{
\item{master_qs_rel}{Path to the master .qs file, relative to the initial working directory.}

\item{orig_setup_rmd}{Path to the Setup_markdown.Rmd file (typically via \code{system.file()}).}

\item{src_paths}{Character vector of source Rmd paths to include in the book.}

\item{book_filename}{Filename (without extension) for the output book.}

\item{output_dir}{Directory where the rendered book will be placed.}

\item{template}{Path to the LaTeX template (.tex) to use for PDF output.}

\item{output_format}{Output format string for \code{render_book}, e.g. "bookdown::pdf_book" or "bookdown::html_book".}

\item{new_session}{Logical; if TRUE, each chapter is rendered in a new R session.}

\item{delete_merged_file}{Logical; passed to Bookdown to delete intermediate merged files.}

\item{root}{Root directory of the project where Bookdown should run (defaults to here::here()).}
}
\description{
This function automates the process of building a Bookdown project with \code{new_session = TRUE}.
It copies and injects setup and chapter files, loads a master .qs environment,
and restores the working directory after completion.
}
\examples{
\dontrun{
build_book(
  master_qs_rel   = "everything_for_bookdown_fixed.qs",
  orig_setup_rmd  = system.file("rmd/Setup_markdown.Rmd", package = "CWP.dataset"),
  src_paths       = c(
    system.file("rmd/index.Rmd", package = "CWP.dataset"),
    system.file("rmd/first_and_first_to_last_and_process.Rmd", package = "CWP.dataset"),
    system.file("rmd/all_child_process.Rmd", package = "CWP.dataset"),
    system.file("rmd/Annexenew.Rmd", package = "CWP.dataset")
  ),
  book_filename      = "tableau_recapbookdowntest",
  output_dir         = "_book",
  template           = system.file("rmd/template.tex", package = "CWP.dataset"),
  output_format      = "bookdown::pdf_document2",
  new_session        = TRUE,
  delete_merged_file = TRUE
)
}
}
