% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_title.R, R/generate_plot.R
\name{cat_title}
\alias{cat_title}
\title{Create a Title with Optional Child Header}
\usage{
cat_title(x, child_headerinside = "")

cat_title(x, child_headerinside = "")
}
\arguments{
\item{x}{A character string representing the title.}

\item{child_headerinside}{An optional character string for the child header.}
}
\value{
A formatted character string.

A formatted character string.
}
\description{
This function formats a title with optional child headers for output.

This function formats a title with optional child headers for output.
}
