% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_temporal_differences.R
\name{compare_temporal_differences}
\alias{compare_temporal_differences}
\title{Calculate and Visualize Temporal Data Differences}
\usage{
compare_temporal_differences(
  parameter_time_dimension,
  init,
  final,
  titre_1,
  titre_2,
  unique_analyse = FALSE
)
}
\arguments{
\item{parameter_time_dimension}{A list of time dimensions to be analyzed.}

\item{init}{Data frame containing initial data.}

\item{final}{Data frame containing final data.}

\item{titre_1}{Title for the first dataset.}

\item{titre_2}{Title for the second dataset.}

\item{unique_analyse}{Logical value indicating whether the analysis is unique.}
}
\value{
A list containing ggplot objects for visualizing the temporal differences.
}
\description{
This function calculates the differences in temporal data between two datasets
and provides visualizations of the differences in percent for each year.
}
\examples{
\dontrun{
compare_temporal_differences(c("Year"), init, final, "Dataset1", "Dataset2", FALSE, "path/to/save")
}
}
