% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_cwp_dataframe_analysis.R
\name{comprehensive_cwp_dataframe_analysis}
\alias{comprehensive_cwp_dataframe_analysis}
\title{This function performs a comprehensive analysis of initial and final datasets, including summary of differences, grouping differences, dimension comparisons, temporal and spatial analyses.}
\usage{
comprehensive_cwp_dataframe_analysis(
  parameter_init,
  parameter_final,
  fig.path = getwd(),
  parameter_fact = "catch",
  parameter_short = FALSE,
  parameter_columns_to_keep = c("Precision", "measurement_unit", "Values dataset 1",
    "Values dataset 2", "Loss / Gain", "Difference (in \%)", "Dimension",
    "Difference in value"),
  parameter_diff_value_or_percent = "Difference (in \%)",
  parameter_UNK_for_not_standards_unit = TRUE,
  parameter_filtering = list(species = NULL, fishing_fleet = NULL),
  parameter_time_dimension = c("time_start"),
  parameter_geographical_dimension = "geographic_identifier",
  parameter_geographical_dimension_groupping = "GRIDTYPE",
  parameter_colnames_to_keep = "all",
  outputonly = FALSE,
  plotting_type = "view",
  print_map = TRUE,
  shapefile_fix = NULL,
  continent = NULL,
  coverage = TRUE,
  parameter_resolution_filter = NULL,
  parameter_titre_dataset_1 = "Dataset 1",
  parameter_titre_dataset_2 = "Dataset 2",
  unique_analyse = FALSE,
  removemap = FALSE,
  topnumber = 6
)
}
\arguments{
\item{fig.path}{Path to save figures.}

\item{parameter_fact}{Fact parameter, default is "#catch".}

\item{parameter_short}{Whether to use short parameter names, default is FALSE.}

\item{parameter_columns_to_keep}{Columns to keep in the comparison.}

\item{parameter_diff_value_or_percent}{Difference value or percent, default is "Difference (in \%)".}

\item{parameter_UNK_for_not_standards_unit}{Whether to use UNK for non-standard units, default is TRUE.}

\item{parameter_filtering}{Filtering parameters.}

\item{parameter_time_dimension}{Time dimension parameters, default is c("time_start").}

\item{parameter_geographical_dimension}{Geographical dimension parameter, default is "geographic_identifier".}

\item{parameter_geographical_dimension_groupping}{Geographical dimension grouping parameter, default is "GRIDTYPE".}

\item{parameter_colnames_to_keep}{Column names to keep in the parameter.}

\item{outputonly}{Whether to output only.}

\item{print_map}{Whether to print the map, default is FALSE.}

\item{coverage}{Whether to analysis the time, geo and other dimension (if no analysing only summary)}

\item{parameter_resolution_filter}{Resolution filter parameter.}

\item{parameter_titre_dataset_1}{Title for dataset 1.}

\item{parameter_titre_dataset_2}{Title for dataset 2.}

\item{unique_analyse}{Whether the analysis is unique.}

\item{topnumber}{Number of first caracteristics of each dimensions to display without groupping}

\item{init}{Initial dataset.}

\item{final}{Final dataset.}
}
\value{
A list containing results of various analyses.
}
\description{
This function performs a comprehensive analysis of initial and final datasets, including summary of differences, grouping differences, dimension comparisons, temporal and spatial analyses.
}
