% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_empreinte_spatiale.R,
%   R/generate_plot.R
\name{fonction_empreinte_spatiale}
\alias{fonction_empreinte_spatiale}
\title{Spatial Footprint Function}
\usage{
fonction_empreinte_spatiale(
  variable_affichee,
  initial_dataset = init,
  final_dataset = final,
  titre_1 = "Dataset 1",
  titre_2 = "Dataset 2",
  shapefile.fix = NULL,
  plotting_type = "plot",
  continent = NULL
)

fonction_empreinte_spatiale(
  variable_affichee,
  initial_dataset = init,
  final_dataset = final,
  titre_1 = "Dataset 1",
  titre_2 = "Dataset 2",
  shapefile.fix = NULL,
  plotting_type = "plot",
  continent = NULL
)
}
\arguments{
\item{variable_affichee}{A character string indicating the measurement unit to be displayed.}

\item{initial_dataset}{A data.table containing the initial measurement data (default: init).}

\item{final_dataset}{A data.table containing the final measurement data (default: final).}

\item{titre_1}{A character string for the title of the first dataset (default: "Dataset 1").}

\item{titre_2}{A character string for the title of the second dataset (default: "Dataset 2").}

\item{shapefile.fix}{A spatial object (sf) for the polygons that defines the geographical areas.}

\item{plotting_type}{A character string indicating the type of plot ("plot" or "view").}

\item{continent}{An optional spatial object for adding continent borders to the plot.}
}
\value{
A plot object representing the spatial footprint of the measurement values.

A plot object representing the spatial footprint of the measurement values.
}
\description{
This function generates a spatial representation of measurement values from two datasets,
allowing for comparison between initial and final datasets using a provided shapefile.

This function generates a spatial representation of measurement values from two datasets,
allowing for comparison between initial and final datasets using a provided shapefile.
}
