% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiple_comparison.R
\name{function_multiple_comparison}
\alias{function_multiple_comparison}
\title{function_multiple_comparison}
\usage{
function_multiple_comparison(
  counting,
  parameter_short,
  sub_list_dir,
  parameters_child_global,
  fig.path,
  coverage = FALSE,
  shapefile.fix,
  continent
)
}
\arguments{
\item{main_dir}{Character. The main directory containing the entities. (jobs/entities)}

\item{connectionDB}{Object. The database connection.}

\item{config}{List. Configuration list containing metadata and options for processing.}

\item{source_authoritylist}{Vector. Vector of source_authority to filter on, "all" being all of them.}

\item{size}{Character string. La taille peut prendre les valeurs suivantes :
\itemize{
\item \code{"long"} (par défaut) : Long with coverage.
\item \code{"middle"} : Long without coverage
\item \code{"short"} : Only first characteristics, first differences and main table of steps
}

function_multiple_comparison <- function(counting, parameter_short, sub_list_dir, parameters_child_global, fig.path, coverage = FALSE, shapefile.fix, continent)}
}
\value{
NULL. The function has side effects, such as writing files and rendering reports.
}
\description{
This function performs various summarizing steps on data related to species and gear types, retrieving data from a database,
processing it, and rendering output reports.
}
