% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_diff.R
\name{geographic_diff}
\alias{geographic_diff}
\title{Geographic Differences}
\usage{
geographic_diff(
  init,
  final,
  shapefile_fix,
  parameter_geographical_dimension,
  parameter_geographical_dimension_groupping,
  continent,
  plotting_type,
  titre_1,
  titre_2,
  outputonly
)
}
\arguments{
\item{init}{Initial dataset}

\item{final}{Final dataset}

\item{shapefile_fix}{Shapefile to be used}

\item{parameter_geographical_dimension}{Parameter for geographical dimension}

\item{parameter_geographical_dimension_groupping}{Parameter for geographical dimension grouping}

\item{continent}{Shapefile of the continent}

\item{plotting_type}{Type of plotting ("plot" or other)}

\item{titre_1}{Title for the first dataset}

\item{titre_2}{Title for the second dataset}

\item{outputonly}{Boolean to specify if output should be saved only}
}
\value{
A list containing the geographic differences and a saved image
}
\description{
This function analyzes geographic data differences between two datasets.
}
