% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-sub-functions-in-the-same-chunk.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate \code{\%in\%} Operator}
\usage{
... \%notin\% NA
}
\arguments{
\item{x}{A vector of elements to check.}

\item{table}{A vector to check against.}
}
\value{
Logical vector indicating if each element of \code{x} is not in \code{table}.
}
\description{
This operator checks if elements are not present in a given vector.
}
