% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R, R/save_image.R
\name{save_image}
\alias{save_image}
\title{Save Plot as Image}
\usage{
save_image(
  title,
  plott = last_plot(),
  folder = NULL,
  fig.pathinside = fig.path
)

save_image(
  title,
  plott = last_plot(),
  folder = NULL,
  fig.pathinside = fig.path
)
}
\arguments{
\item{title}{A character string representing the title of the plot.}

\item{plott}{The plot object to save.}

\item{folder}{The folder where the image will be saved.}

\item{fig.pathinside}{The path for saving the figure.}

\item{find_and_print}{Logical indicating if results should be printed.}
}
\value{
None

None
}
\description{
This function saves the current plot as an image in a specified folder.

This function saves the current plot as an image in a specified folder.
}
