% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coverage_analysis.R
\name{spatial_coverage_analysis}
\alias{spatial_coverage_analysis}
\title{Calculate and Visualize Spatial Coverage}
\usage{
spatial_coverage_analysis(
  init,
  final,
  titre_1 = "Dataset 1",
  titre_2 = "Dataset 2",
  shapefile.fix,
  plotting_type,
  continent,
  print_map = TRUE,
  Grouppedgridtype
)
}
\arguments{
\item{init}{Initial dataset.}

\item{final}{Final dataset.}

\item{titre_1}{Title for the first dataset.}

\item{titre_2}{Title for the second dataset.}

\item{shapefile.fix}{Shapefile for fixing spatial data.}

\item{plotting_type}{Type of plotting to be used.}

\item{continent}{Data frame of continent shapes for reference.}

\item{print_map}{Logical indicating whether to print the map.}

\item{Grouppedgridtype}{Data frame containing grouped gridtype data.}
}
\value{
A list containing the spatial coverage maps and related information.
}
\description{
This function calculates the spatial coverage for different units and provides visualizations
of the values over spatial regions for each dataset.
}
\examples{
\dontrun{
spatial_coverage_analysis(units, init, final, "Dataset1", "Dataset2", shapefile.fix, "plot", continent, TRUE, Grouppedgridtype, "path/to/save")
}
}
\author{
Bastien Grasset, \email{bastien.grasset@ird.fr}
}
