% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/strata_in_georef_but_not_in_nominal_report_launching.R
\name{strata_in_georef_but_not_in_nominal_report_launching}
\alias{strata_in_georef_but_not_in_nominal_report_launching}
\title{Launch Strata in Georef but not in Nominal Report}
\usage{
strata_in_georef_but_not_in_nominal_report_launching(
  main.dir,
  connectionDB,
  uploadgoogledrive = TRUE
)
}
\arguments{
\item{main.dir}{Directory containing the main data files.}

\item{connectionDB}{Database connection object.}

\item{uploadgoogledrive}{Logical indicating whether to upload the report to Google Drive.}
}
\value{
Data frame containing the upgraded nominal data.
}
\description{
This function launches a report comparing strata in georef data but not in nominal data.
}
